requirements.txt
keras==2.13.0
numpy==1.24.3
Pillow==9.5.0
scikit_learn==1.3.1
streamlit==1.27.1
tensorflow==2.13.0
